Fix long-encoding under Python 3
Sometimes I forget that Python 3 got rid of the `long` type — it is
just `int` now. This commit creates the `long` and `unicode` types
in python 3 as type aliases.

+++ neet/statespace.py
@@ -1,6 +1,7 @@
 # Copyright 2016-2017 ELIFE. All rights reserved.
 # Use of this source code is governed by a MIT
 # license that can be found in the LICENSE file.
+from .python3 import *
 import numpy as np
 
 class StateSpace(object):
@@ -213,7 +214,7 @@ class StateSpace(object):
                 encoded += place * x
                 place *= b
 
-        return encoded
+        return long(encoded)
 
     def encode(self, state):
         """
